dashboard/python/requirements.txt
absl-py==1.4.0
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
anywidget==0.9.18
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
boto3==1.38.0
botocore==1.38.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chatlas @ git+https://github.com/posit-dev/chatlas.git@a31161995216335e478f3aff80a71f53596fa0ff
chevron==0.14.0
click==8.1.8
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
duckdb==1.2.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.14.0
htmltools==0.6.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kiwisolver==1.4.8
linkify-it-py==2.0.3
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.2
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook-shim==0.2.4
numpy==1.26.4
nvidia-nccl-cu12==2.26.2.post1
openai==1.70.0
orjson==3.10.16
overrides==7.7.0
packaging==25.0
palmerpenguins==0.1.4
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.7
plotly==6.0.1
polars==1.27.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psygnal==0.12.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.2
pydantic-core==2.33.1
pygments==2.19.1
pyjwt==2.10.1
pyopenssl==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
querychat @ git+https://github.com/posit-dev/querychat.git@a218fb914963a4477598c8f4d0081bae043de286#subdirectory=python-package
questionary==2.1.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsconnect-python==1.25.2
s3fs==2025.3.2
s3transfer==0.12.0
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
send2trash==1.8.3
setuptools==78.1.0
shiny @ git+https://github.com/posit-dev/py-shiny.git@58e82de55683435c8f7ab852978cf454131cd5ab
shinywidgets==0.5.2
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.14.1
snowflake-ml-python==1.8.1
snowflake-snowpark-python==1.30.0
snowflake-sqlalchemy==1.7.3
sortedcontainers==2.4.0
soupsieve==2.6
sqlalchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==2.1.4
yarl==1.19.0
